requirements.txt
# API server
fastapi
uvicorn

# Core dependencies
python-dotenv
pydantic
psycopg2-binary
requests

# LangChain / LangGraph
langchain-core
langchain-google-genai
langchain-huggingface
sentence-transformers
langchain-tavily
langgraph

# Knowledge Graph (Graphiti + Neo4j)
graphiti-core[google-genai]

# Browser automation (JS-rendered page crawling: Naver Cafe/Blog etc.)
playwright

# Telegram bot
aiogram>=3.0,<4.0
anthropic
croniter
